modules/grpc-client-frontend/requirements.txt
flask
flask-restful
grpcio
grpcio-tools

